Burst App
This is an in-progress iOS app for a strategic, trick-based card game
called Burst.  It is fully playable, both with multiple human players
and with computer players.  However, currently the AI does not play
very well, and I am working on algorithms to make it smarter.  Till
then, here is the rudimentary game: it generates graphics, responds to
touch events, animates card motion, and understands/incorporates the
rules of the game.